backend/requirements.txt
requests

# no system packages
python-multipart
fastapi
uvicorn
lz4
pycryptodome